requirements.txt
aiofiles==24.1.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
apsw==3.36.0.post1
apturl==0.5.2
asgiref==3.8.1
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.1.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.2.1
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2
django-debug-toolbar==6.0.0
dnspython==2.2.1
duplicity==0.8.21
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
fasteners==0.14.1
feedparser==6.0.8
ffmpy==0.6.1
filelock==3.19.1
Flask==3.1.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.2
Flask-Migrate==4.1.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
Flask-WTF==1.0.1
fonttools==4.59.2
fs==2.4.12
fsspec==2025.9.0
future==0.18.2
gast==0.5.2
gradio==5.44.1
gradio_client==1.12.1
greenlet==1.1.3
groovy==0.1.2
gyp==0.1
h11==0.16.0
hf-xet==1.1.9
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.6.4
ipython==7.31.1
itsdangerous==2.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.2
kagglehub==0.3.13
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.3
mdurl==0.1.2
mechanize==0.4.7
mixin==1.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
mysql-connector-python==9.4.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
orjson==3.11.3
packaging==25.0
pandas==2.3.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2025.0.4
platformdirs==4.4.0
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
ptyprocess==0.7.0
py7zr==0.11.3+dfsg
pycairo==1.20.1
pychm==0.8.6
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.14.0
PyMySQL==1.1.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.9.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.1.1
python-multipart==0.0.20
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2021.11.10
reportlab==3.6.8
repoze.lru==0.7
requests==2.25.1
rich==14.1.0
Routes==2.5.1
RPi.GPIO==0.7.1
ruff==0.12.11
safehttpx==0.1.6
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
SecretStorage==3.3.1
seeborn==0.4.3
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.17.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.3
sympy==1.9
systemd-python==234
texttable==1.6.4
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.1.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.35.0
virtualenv==20.34.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.0.1
xdg==5
xkit==0.0.0
zeroconf==0.38.3
zipp==1.0.0
